+++ finam/exception.py
@@ -1,0 +1,31 @@
+__all__ = ['FinamExportError',
+           'FinamDownloadError',
+           'FinamThrottlingError',
+           'FinamParsingError',
+           'FinamObjectNotFoundError',
+           'FinamTooLongTimeframeError'
+           ]
+
+
+class FinamExportError(Exception):
+    pass
+
+
+class FinamDownloadError(FinamExportError):
+    pass
+
+
+class FinamThrottlingError(FinamExportError):
+    pass
+
+
+class FinamParsingError(FinamExportError):
+    pass
+
+
+class FinamObjectNotFoundError(FinamExportError):
+    pass
+
+
+class FinamTooLongTimeframeError(FinamExportError):
+    pass