+++ samples/lookups.py
@@ -2,7 +2,7 @@
 # coding: utf-8
 import logging
 
-from finam.export import Exporter, Market, LookupComparator
+from finam import Exporter, Market, LookupComparator
 
 """
 This sample script shows off lookup capabilities